move test code to tests/backends
